+++ Sources/ReplicantSwift/ReplicantConfigTemplate.swift
@@ -9,13 +9,13 @@ import Foundation
 
 public struct ReplicantConfigTemplate: Codable
 {
-    public var chunkSize: Int
+    public var chunkSize: UInt16
     public var chunkTimeout: Int
     public var addSequences: [SequenceModel]?
     public var removeSequences: [SequenceModel]?
     
     
-    public init?(chunkSize: Int, chunkTimeout: Int, addSequences: [SequenceModel]?, removeSequences: [SequenceModel]?)
+    public init?(chunkSize: UInt16, chunkTimeout: Int, addSequences: [SequenceModel]?, removeSequences: [SequenceModel]?)
     {
         guard chunkSize >= keySize + aesOverheadSize
             else
