+++ Sources/ReplicantSwift/ReplicantServerConfig.swift
@@ -9,12 +9,12 @@ import Foundation
 
 public struct ReplicantServerConfig: Codable
 {
-    public var chunkSize: Int
+    public var chunkSize: UInt16
     public var chunkTimeout: Int
     public var addSequences: [SequenceModel]?
     public var removeSequences: [SequenceModel]?
     
-    public init?(serverPublicKey: SecKey, chunkSize: Int, chunkTimeout: Int, addSequences: [SequenceModel]?, removeSequences: [SequenceModel]?)
+    public init?(serverPublicKey: SecKey, chunkSize: UInt16, chunkTimeout: Int, addSequences: [SequenceModel]?, removeSequences: [SequenceModel]?)
     {
         guard chunkSize >= keySize + aesOverheadSize
             else
